+++ src/main/java/it/euris/academy/centrosportivo/repository/CourseRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigInteger;
 
-public interface CourseRepository extends JpaRepository<Course, BigInteger> {
+public interface CourseRepository extends JpaRepository<Course, Long> {
 }
