+++ src/main/java/it/euris/academy/centrosportivo/repository/CustomerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigInteger;
 
-public interface CustomerRepository extends JpaRepository<Customer, BigInteger> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
 }
